requirements.txt
google-cloud-pubsub>=2.27.0
google-cloud-storage>=2.19.0
apache-beam[gcp]>=2.61.0
google-cloud-bigquery>=3.27.0
google-cloud-aiplatform>=1.75.0
pandas>=2.2.3
google-auth>=2.37.0
google-generativeai>=0.8.3
matplotlib>=3.10.0
streamlit>=1.41.1
ratelimit>=2.2.1
ipython>=8.31.0
google-cloud-texttospeech[beta]>=2.23.0
firebase-admin>=6.6.0
pydub>=0.25.1
pymongo[srv]>=4.10.1
langchain-google-alloydb-pg>=0.9.0
langchain-google-vertexai>=2.0.11
pytest>=7.3.1
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
starlette>=0.45.2
cryptography>=44.0.0
google-genai>=0.6.0
google-cloud-logging>=3.11.3
google-cloud-secret-manager>=2.22.1
protobuf>=3.20.0
requests>=2.32.0
google-cloud-firestore>=2.20.0
langchain-google-firestore>=0.5.0
pytz>=2024.2
google-cloud-video-transcoder>=1.14.0